ep-cli/requirements.txt
wheel
cbor2==4.1.2
pika==1.1.0
PyYAML==5.1.2
tabulate==0.8.*
